pyproject.toml
[project]
name = "swc-url-shortener"
version = "0.1.0"
description = "SWC URL Shortener Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-stubs[compatible-mypy]>=6.0.2",
    "djangorestframework>=3.17.1",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "djangorestframework-stubs[compatible-mypy]>=3.16.9",
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
plugins = [
    "mypy_django_plugin.main", 
    "mypy_drf_plugin.main",
]

[tool.django-stubs]
django_settings_module = "swc.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "swc.settings"
python_files = ["test_*.py"]
